// src/api/generated/types.gen.ts
export type Booking = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    userId?: number;
    user?: User;
    startDate?: string;
    endDate?: string | null;
    recurrence?: number | null;
    reservations?: Array<Reservation> | null;
    calendarInviteUid?: string | null;
};
// ...
export type Delegation = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    userId?: number;
    user?: User;
    delegateId?: number;
    delegate?: User;
    email?: string | null;
};

export type Division = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    name?: string | null;
    adminEmail?: string | null;
    floors?: Array<FloorDivision> | null;
    users?: Array<User> | null;
};

export type Favourite = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    workstationId?: number;
    workstation?: Workstation;
    userId?: number;
    user?: User;
};
// ...
export type FloorDivision = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    divisionId?: number;
    floorId?: number;
    division?: Division;
    floor?: Floor;
};

export type FloorPlan = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    floorPlanData?: string | null;
    fileType?: string | null;
    floorPlanUrl?: string | null;
    fileName?: string | null;
    floorId?: number;
    floor?: Floor;
};

export type NotificationType = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    value?: string | null;
    users?: Array<User> | null;
};

export type ReportFilter = {
    buildingId?: number | null;
    floorIds?: Array<number> | null;
    startDate?: string;
    endDate?: string;
    email?: string | null;
    divisionIds?: Array<number> | null;
    tableName?: string | null;
    searchFilter?: string | null;
};

export type Reservation = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    bookingId?: number;
    workstationId?: number;
    date?: string;
    reservationTimeSlotId?: number;
    booking?: Booking;
    workstation?: Workstation;
    reservationTimeSlot?: ReservationTimeSlot;
    reservationStatuses?: Array<ReservationStatusHistory> | null;
};
// ...
export type RoleType = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    name?: string | null;
    usersRoles?: Array<UserRole> | null;
};

export type User = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    emailAddress?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    displayName?: string | null;
    managerEmail?: string | null;
    phoneNumber?: string | null;
    postalCode?: string | null;
    notificationTypeId?: number;
    userStatusId?: number;
    divisionId?: number | null;
    isManager?: boolean;
    notificationType?: NotificationType;
    userStatus?: UserStatus;
    division?: Division;
    bookings?: Array<Booking> | null;
    userRoles?: Array<UserRole> | null;
    favourites?: Array<Favourite> | null;
    delegations?: Array<Delegation> | null;
};

export type UserRole = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    userId?: number;
    user?: User;
    roleTypeId?: number;
    roleType?: RoleType;
};

export type UserStatus = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    name?: string | null;
};

export type Workstation = {
    lastModifiedDate?: string;
    lastModifiedBy?: string | null;
    createDate?: string;
    createdBy?: string | null;
    isDeleted?: boolean;
    id?: number;
    number?: string | null;
    note?: string | null;
    isEnabled?: boolean;
    floorId?: number;
    disabledPersonWorkspace?: boolean;
    floor?: Floor;
    reservations?: Array<Reservation> | null;
    attributes?: Array<WorkstationAttribute> | null;
    reviews?: Array<WorkstationReview> | null;
    workstationClosures?: Array<WorkstationClosure> | null;
    workstationIssues?: Array<WorkstationIssue> | null;
    userFavourites?: Array<Favourite> | null;
};